inductor(CPU): add ISA check before do cpu fx packed weight
[ghstack-poisoned]
