+++ aten/src/ATen/native/mkldnn/RegisterMkldnnOpContextClass.cpp
@@ -6,11 +6,16 @@
 
 #if AT_MKLDNN_ENABLED()
 
+#include <ATen/native/mkldnn/Utils.h>
 namespace at {
 namespace native {
 namespace mkldnn {
 
 using namespace internal::convolution;
+
+bool is_mkldnn_bf16_supported() {
+  return mkldnn_bf16_device_check();
+}
 
 TORCH_LIBRARY(mkldnn, m) {
   m.class_<ConvOpContext>(TORCH_SELECTIVE_CLASS("ConvOpContext"))
@@ -48,6 +53,7 @@ TORCH_LIBRARY(mkldnn, m) {
       "mkldnn::_reorder_convolution_transpose_weight(Tensor self, int[2] padding=0, int[2] output_padding=0, int[2] stride=1, int[2] dilation=1, int groups=1, int[]? input_size=None) -> Tensor Y"));
   m.def(TORCH_SELECTIVE_SCHEMA(
       "mkldnn::_reorder_linear_weight(Tensor self, int? batch_size=None) -> Tensor Y"));
+  m.def("_is_mkldnn_bf16_supported", &is_mkldnn_bf16_supported);
 }
 
 TORCH_LIBRARY(mkldnn_prepacked, m) {
